Disable enter until modal *fully* opens
Issue #17
The showModal status isn’t flipped until the modal is fully opened or
closed.  The transition introduces a brief time period that the modal
isn’t fully opened.  The enter key is clicked and the modal closes.  It
appears as if the modal disappears immediately.

Aside: Clicking area outside of the modal also closes the modal as well.